requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.12
numpy>=1.20.0
Pillow>=9.0.0
pyyaml>=6.0
yacs>=0.1.8
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.8.0
opencv-python>=4.5.0
